REF: hack to be able to share distutils utilities.
We took the ugly trick from NumPy to separate build vs normal use of
sklearn imports.
