+++ sklearn/build_utils/__init__.py
@@ -1,0 +1,1 @@
+from .blas import get_blas_info
