+++ sklearn/setup.py
@@ -3,32 +3,6 @@ from os.path import join
 import warnings
 
 
-def get_blas_info():
-    from numpy.distutils.system_info import get_info
-
-    def atlas_not_found(blas_info_):
-        def_macros = blas_info.get('define_macros', [])
-        for x in def_macros:
-            if x[0] == "NO_ATLAS_INFO":
-                # if x[1] != 1 we should have lapack
-                # how do we do that now?
-                return True
-            if x[0] == "ATLAS_INFO":
-                if "None" in x[1]:
-                    # this one turned up on FreeBSD
-                    return True
-        return False
-
-    blas_info = get_info('blas_opt', 0)
-    if (not blas_info) or atlas_not_found(blas_info):
-        cblas_libs = ['cblas']
-        blas_info.pop('libraries', None)
-    else:
-        cblas_libs = blas_info.pop('libraries', [])
-
-    return cblas_libs, blas_info
-
-
 def configuration(parent_package='', top_path=None):
     from numpy.distutils.misc_util import Configuration
     from numpy.distutils.system_info import get_info, BlasNotFoundError
@@ -41,6 +15,7 @@ def configuration(parent_package='', top_path=None):
     config = Configuration('sklearn', parent_package, top_path)
 
     config.add_subpackage('__check_build')
+    config.add_subpackage('build_utils')
     config.add_subpackage('svm')
     config.add_subpackage('datasets')
     config.add_subpackage('datasets/tests')
